Restructure the release storage to a phased system
This commit replaces the separate hashed releases and rsync upload
areas with an area split into phases: candidate-draft for files being
prepared for vote, candidate-release for files being voted upon,
distributable-draft for files in the process of being finalised for
release, and distributable-release for approved and distributed
files. This commit updates all of the related functionality, including
uploads via browser and rsync, downloads, and file management. The UI
has also changed to reflect the new terminology. There is still some
hybridisation of the interface, and deprecated code.