Update tutorial.html minor text edits and questions
line 21 - If this is a tutorial, let's use the full name for the service at the start, and the TLA  for it later
line 58 - avoid passive voice.
line 58 - 'chosen mailing list' presents a problem to the user of the tutorial. Who chooses it, how, and when in the process?
line 67 - change 'report' to 'vote'

I note the use of 'use' in relation to clicking a button. Would not 'click' be clearer, unless there is some other behavior that is possible?